+++ lib/model/wellness/Appointment.dart
@@ -95,19 +95,14 @@ class Appointment with Explore, Favorite {
   }
 
   String? get imageKeyBasedOnCategory { //Keep consistent images
-      String? toutImageUrl;
-      switch (type) {
-        case AppointmentType.in_person:
-          toutImageUrl = 'photo-building';
-          break;
-        case AppointmentType.online:
-          toutImageUrl = 'photo-online';
-          break;
-        default:
-          toutImageUrl = imageUrl!;
-          break;
-      }
-      return toutImageUrl;
+    return (type != null) ? imageKeyForAppointmentType(type!) : imageUrl;
+  }
+
+  static String imageKeyForAppointmentType(AppointmentType appointmentType) {
+    switch (appointmentType) {
+      case AppointmentType.in_person: return 'photo-building';
+      case AppointmentType.online: return 'photo-online';
+    }
   }
 
   static List<Appointment>? listFromJson(List<dynamic>? jsonList) {
@@ -268,6 +263,8 @@ class AppointmentLocation {
 
   AppointmentLocation({this.id, this.latitude, this.longitude, this.title, this.phone});
 
+  String? get address => title; //TBD: ?
+
   Map<String, dynamic> toJson() {
     return {
       'id': id,
@@ -380,8 +377,14 @@ class AppointmentsAccount {
 class AppointmentProvider {
   final String? id;
   final String? name;
+  final bool? supportsSchedule;
+  final bool? supportsReschedule;
+  final bool? supportsCancel;
 
-  AppointmentProvider({this.id, this.name});
+  AppointmentProvider({
+    this.id, this.name,
+    this.supportsSchedule, this.supportsReschedule, this.supportsCancel
+  });
 
   // JSON Serialization
 
@@ -389,6 +392,9 @@ class AppointmentProvider {
     return (json != null) ? AppointmentProvider(
       id: JsonUtils.stringValue(json['id']),
       name: JsonUtils.stringValue(json['name']),
+      supportsSchedule: JsonUtils.boolValue(json['supports_schedule']),
+      supportsReschedule: JsonUtils.boolValue(json['supports_reschedule']),
+      supportsCancel: JsonUtils.boolValue(json['supports_cancel']),
     ) : null;
   }
 
@@ -396,6 +402,9 @@ class AppointmentProvider {
     return {
       'id': id,
       'name': name,
+      'supports_schedule': supportsSchedule,
+      'supports_reschedule': supportsReschedule,
+      'supports_cancel': supportsCancel,
     };
   }
 
@@ -427,12 +436,18 @@ class AppointmentProvider {
   bool operator==(dynamic other) =>
     (other is AppointmentProvider) &&
     (id == other.id) &&
-    (name == other.name);
+    (name == other.name) &&
+    (supportsSchedule == other.supportsSchedule) &&
+    (supportsReschedule == other.supportsReschedule) &&
+    (supportsCancel == other.supportsCancel);
 
   @override
   int get hashCode =>
     (id?.hashCode ?? 0) ^
-    (name?.hashCode ?? 0);
+    (name?.hashCode ?? 0) ^
+    (supportsSchedule?.hashCode ?? 0) ^
+    (supportsReschedule?.hashCode ?? 0) ^
+    (supportsCancel?.hashCode ?? 0);
 
   // Accessories