+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Schedule appointment preview progress. [#3074](https://github.com/rokwire/illinois-app/issues/3074).
+
 ## [4.3.36] - 2023-04-05
 ### Changed
 - Display mental health resources grouped by sections [#3084](https://github.com/rokwire/illinois-app/issues/3084).